Combined example utils into header-only files
